perf: optimised MapUtil with safer implementations
