+++ settings.gradle.kts
@@ -28,3 +28,4 @@ include(
     "simple-jwt-spring-boot-starter",
     "property-guard-spring-boot-starter"
 )
+include("map-util-unsafe")
